+++ plugins/nf-schema/src/main/nextflow/validation/SchemaValidator.groovy
@@ -119,6 +119,12 @@ class SchemaValidator extends PluginExtensionPoint {
     // The length of the terminal
     private Integer terminalLength = System.getenv("COLUMNS")?.toInteger() ?: 100
 
+    // The configuration class
+    private ValidationConfig config
+
+    // The session
+    private Session session
+
     @Override
     protected void init(Session session) {
         def plugins = session?.config?.navigate("plugins") as ArrayList
@@ -140,14 +146,31 @@ class SchemaValidator extends PluginExtensionPoint {
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
             """)
         }
-    }
 
-    Session getSession(){
-        Global.getSession() as Session
-    }  
+        this.session = session
 
-    ValidationConfig getConfig() {
-        new ValidationConfig(session.config.navigate('validation') as Map)
+        // Help message logic
+        def Map params = (Map)session.params ?: [:]
+        config = new ValidationConfig(session?.config?.navigate('validation') as Map, params)
+        def Boolean containsFullParameter = params.containsKey(config.help.fullParameter) && params[config.help.fullParameter]
+        def Boolean containsShortParameter = params.containsKey(config.help.shortParameter) && params[config.help.shortParameter]
+        if (config.help.enabled && (containsFullParameter || containsShortParameter)) {
+            def String help = ""
+            def HelpMessage helpMessage = new HelpMessage(config, session)
+            help += helpMessage.getBeforeText()
+            if (containsFullParameter) {
+                log.debug("Printing out the full help message")
+                help += helpMessage.getFullHelpMessage()
+            } else if (containsShortParameter) {
+                log.debug("Printing out the short help message")
+                def paramValue = params.get(config.help.shortParameter)
+                help += helpMessage.getShortHelpMessage(paramValue instanceof String ? paramValue : "")
+            }
+            help += helpMessage.getAfterText()
+            log.info(help)
+            System.exit(0)
+        }
+
     }
 
     boolean hasErrors() { errors.size()>0 }
@@ -238,7 +261,7 @@ class SchemaValidator extends PluginExtensionPoint {
 
         def Map params = initialiseExpectedParams(session.params)
         def String baseDir = session.baseDir.toString()
-        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : 'nextflow_schema.json'
+        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : config.parametersSchema
         log.debug "Starting parameters validation"
 
         // Clean the parameters
@@ -333,154 +356,41 @@ class SchemaValidator extends PluginExtensionPoint {
         return new Tuple (expectedParams, enums)
     }
 
-
-    //
-    // Wrap too long text
-    //
-    String wrapText(String text) {
-        List olines = []
-        String oline = ""
-        text.split(" ").each() { wrd ->
-            if ((oline.size() + wrd.size()) <= terminalLength) {
-                oline += wrd + " "
-            } else {
-                olines += oline
-                oline = wrd + " "
-            }
-        }
-        olines += oline
-        return olines.join("\n")
-    }
-
     //
     // Beautify parameters for --help
     //
     @Function
-    String paramsHelp(
-        Map options = null,
+    public String paramsHelp(
+        Map options = [:],
         String command
     ) {
-        def Map params = initialiseExpectedParams(session.params)
+        // TODO add link to help message migration guide once created
+        if (!options.containsKey("hideWarning") || options.hideWarning == false) {
+            log.warn("""
+Using `paramsHelp()` is not recommended. Check out the help message migration guide: <url>
+If you intended to use this function, please add the following option to the input of the function:
+    `hideWarning: true`
 
-        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : 'nextflow_schema.json'
-
-        def colors = Utils.logColours(config.monochromeLogs)
-        Integer num_hidden = 0
-        String output  = ''
-        output        += 'Typical pipeline command:\n\n'
-        output        += "  ${colors.cyan}${command}${colors.reset}\n\n"
-        Map paramsMap = paramsLoad( Path.of(Utils.getSchemaPath(session.baseDir.toString(), schemaFilename)) )
-        Integer maxChars  = paramsMaxChars(paramsMap) + 1
-
-        // Make sure the hidden parameters count is 0
-        hiddenParametersCount = 0
-
-        // If a value is passed to help
-        if (params.help instanceof String) {
-            def String paramName = params.help
-            def List<String> paramNames = params.help.tokenize(".") as List<String>
-            def Map paramOptions = [:]
-            for (group in paramsMap.keySet()) {
-                def Map group_params = paramsMap.get(group) as Map // This gets the parameters of that particular group
-                if (group_params.containsKey(paramNames[0])) {
-                    paramOptions = group_params.get(paramNames[0]) as Map 
-                }
-            }
-            if (paramNames.size() > 1) {
-                paramNames.remove(0)
-                paramNames.each {
-                    paramOptions = (Map) paramOptions?.properties?[it] ?: [:]
-                }
-            }
-            if (!paramOptions) {
-                throw new Exception("Specified param '${paramName}' does not exist in JSON schema.")
-            }
-            output += getDetailedHelpString(paramName, paramOptions, colors)
-            output += "-${colors.dim}----------------------------------------------------${colors.reset}-"
-            return output
+Please contact the pipeline maintainer(s) if you see this warning as a user.
+            """)
         }
 
-        for (group in paramsMap.keySet()) {
-            Integer num_params = 0
-            String group_output = "$colors.underlined$colors.bold$group$colors.reset\n"
-            def Map group_params = paramsMap.get(group) as Map // This gets the parameters of that particular group
-            group_output += getHelpList(group_params, colors, maxChars).join("\n") + "\n"
-            if (group_output != "\n"){
-                output += group_output
-            }
-        }
-        if (num_hidden > 0){
-            output += "$colors.dim !! Hiding $num_hidden params, use the 'validation.showHiddenParams' config value to show them !!\n$colors.reset"
-        }
-        output += "-${colors.dim}----------------------------------------------------${colors.reset}-"
-        return output
-    }
-
-    //
-    // Get help text in string format
-    //
-    private List<String> getHelpList(Map<String,Map> params, Map colors, Integer maxChars, String parentParameter = "") {
-        def List helpMessage = []
-        for (String paramName in params.keySet()) {
-            def Map paramOptions = params.get(paramName) as Map 
-            if (paramOptions.hidden && !config.showHiddenParams) {
-                hiddenParametersCount += 1
-                continue
-            }
-            def String type = '[' + paramOptions.type + ']'
-            def String enumsString = ""
-            if (paramOptions.enum != null) {
-                def List enums = (List) paramOptions.enum
-                def String chopEnums = enums.join(", ")
-                if(chopEnums.length() > terminalLength){
-                    chopEnums = chopEnums.substring(0, terminalLength-5)
-                    chopEnums = chopEnums.substring(0, chopEnums.lastIndexOf(",")) + ", ..."
-                }
-                enumsString = " (accepted: " + chopEnums + ") "
-            }
-            def String description = paramOptions.description ? paramOptions.description as String + " " : ""
-            def defaultValue = paramOptions.default != null ? "[default: " + paramOptions.default.toString() + "] " : ''
-            def String nestedParamName = parentParameter ? parentParameter + "." + paramName : paramName
-            def String nestedString = paramOptions.properties ? "(This parameter has sub-parameters. Use '--help ${nestedParamName}' to see all sub-parameters) " : ""
-            def descriptionDefault = description + colors.dim + enumsString + defaultValue + colors.reset + nestedString
-            // Wrap long description texts
-            // Loosely based on https://dzone.com/articles/groovy-plain-text-word-wrap
-            if (descriptionDefault.length() > terminalLength){
-                descriptionDefault = wrapText(descriptionDefault)
-            }
-            helpMessage.add("  --" +  paramName.padRight(maxChars) + colors.dim + type.padRight(10) + colors.reset + descriptionDefault)
-        }
-        return helpMessage
-    }
-
-    //
-    // Get a detailed help string from one parameter
-    //
-    private String getDetailedHelpString(String paramName, Map paramOptions, Map colors) {
-        def String helpMessage = "--" + paramName + '\n'
-        for (option in paramOptions) {
-            def String key = option.key
-            if (key == "fa_icon" || (key == "type" && option.value == "object")) {
-                continue
-            }
-            if (key == "properties") {
-                def Map subParamsOptions = option.value as Map
-                def Integer maxChars = paramsMaxChars(subParamsOptions) + 2
-                def String subParamsHelpString = getHelpList(subParamsOptions, colors, maxChars, paramName)
-                    .collect {
-                        "      ." + it[4..it.length()-1]
-                    }
-                    .join("\n")
-                helpMessage += "    " + colors.dim + "options".padRight(11) + ": " + colors.reset + "\n" + subParamsHelpString + "\n"
-                continue
-            }
-            def String value = option.value
-            if (value.length() > terminalLength) {
-                value = wrapText(value)
-            }
-            helpMessage += "    " + colors.dim + key.padRight(11) + ": " + colors.reset + value + '\n'
-        }
-        return helpMessage
+        def Map params = session.params
+        def Map validationConfig = (Map)session.config.navigate("validation")
+        validationConfig.parametersSchema = options.containsKey('parameters_schema') ? options.parameters_schema as String : config.parametersSchema
+        validationConfig.help = (Map)validationConfig.help + [command: command, beforeText: "", afterText: ""]
+        def ValidationConfig copyConfig = new ValidationConfig(validationConfig, params)
+        def HelpMessage helpMessage = new HelpMessage(copyConfig, session)
+        def String help = helpMessage.getBeforeText()
+        def List<String> helpBodyLines = helpMessage.getShortHelpMessage(params.help && params.help instanceof String ? params.help : "").readLines()
+        help += helpBodyLines.findAll {
+            // Remove added ungrouped help parameters
+            !it.startsWith("--${copyConfig.help.shortParameter}") && 
+            !it.startsWith("--${copyConfig.help.fullParameter}") && 
+            !it.startsWith("--${copyConfig.help.showHiddenParameter}")
+        }.join("\n")
+        help += helpMessage.getAfterText()
+        return help
     }
 
     //
@@ -492,7 +402,7 @@ class SchemaValidator extends PluginExtensionPoint {
         WorkflowMetadata workflow
         ) {
         
-        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : 'nextflow_schema.json'
+        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : config.parametersSchema
         def Map params = session.params
         
         // Get a selection of core Nextflow workflow options
@@ -517,12 +427,13 @@ class SchemaValidator extends PluginExtensionPoint {
 
         // Get pipeline parameters defined in JSON Schema
         def Map paramsSummary = [:]
-        def Map paramsMap = paramsLoad( Path.of(Utils.getSchemaPath(session.baseDir.toString(), schemaFilename)) )
+        def Map paramsMap = Utils.paramsLoad( Path.of(Utils.getSchemaPath(session.baseDir.toString(), schemaFilename)) )
         for (group in paramsMap.keySet()) {
             def Map groupSummary = getSummaryMapFromParams(params, paramsMap.get(group) as Map)
             paramsSummary.put(group, groupSummary)
         }
-        return [ 'Core Nextflow options' : workflowSummary ] << paramsSummary as LinkedHashMap
+        paramsSummary.put('Core Nextflow options', workflowSummary)
+        return paramsSummary
     }
 
 
@@ -585,7 +496,7 @@ class SchemaValidator extends PluginExtensionPoint {
 
         def Map params = session.params
 
-        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : 'nextflow_schema.json'
+        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : config.parametersSchema
 
         def colors = Utils.logColours(config.monochromeLogs)
         String output  = ''
@@ -593,7 +504,7 @@ class SchemaValidator extends PluginExtensionPoint {
         paramsMap.each { key, value ->
             paramsMap[key] = flattenNestedParamsMap(value as Map)
         }
-        def maxChars  = paramsMaxChars(paramsMap)
+        def maxChars  = Utils.paramsMaxChars(paramsMap)
         for (group in paramsMap.keySet()) {
             def Map group_params = paramsMap.get(group) as Map // This gets the parameters of that particular group
             if (group_params) {
@@ -653,100 +564,5 @@ class SchemaValidator extends PluginExtensionPoint {
             }
         }
         return new_params
-    }
-
-    //
-    // This function tries to read a JSON params file
-    //
-    private static LinkedHashMap paramsLoad(Path json_schema) {
-        def paramsMap = new LinkedHashMap()
-        try {
-            paramsMap = paramsRead(json_schema)
-        } catch (Exception e) {
-            println "Could not read parameters settings from JSON. $e"
-            paramsMap = new LinkedHashMap()
-        }
-        return paramsMap
-    }
-
-    //
-    // Method to actually read in JSON file using Groovy.
-    // Group (as Key), values are all parameters
-    //    - Parameter1 as Key, Description as Value
-    //    - Parameter2 as Key, Description as Value
-    //    ....
-    // Group
-    //    -
-    private static LinkedHashMap paramsRead(Path json_schema) throws Exception {
-        def slurper = new JsonSlurper()
-        def Map schema = (Map) slurper.parse( json_schema )
-        // $defs is the adviced keyword for definitions. Keeping defs in for backwards compatibility
-        def Map schema_defs = (Map) (schema.get('$defs') ?: schema.get("defs"))
-        def Map schema_properties = (Map) schema.get('properties')
-        /* Tree looks like this in nf-core schema
-        * $defs <- this is what the first get('$defs') gets us
-                group 1
-                    title
-                    description
-                        properties
-                        parameter 1
-                            type
-                            description
-                        parameter 2
-                            type
-                            description
-                group 2
-                    title
-                    description
-                        properties
-                        parameter 1
-                            type
-                            description
-        * properties <- parameters can also be ungrouped, outside of $defs
-                parameter 1
-                    type
-                    description
-        */
-
-        def paramsMap = new LinkedHashMap()
-        // Grouped params
-        if (schema_defs) {
-            for (group in schema_defs) {
-                def Map group_property = (Map) group.value['properties'] // Gets the property object of the group
-                def String title = (String) group.value['title']
-                def sub_params = new LinkedHashMap()
-                group_property.each { innerkey, value ->
-                    sub_params.put(innerkey, value)
-                }
-                paramsMap.put(title, sub_params)
-            }
-        }
-
-        // Ungrouped params
-        if (schema_properties) {
-            def ungrouped_params = new LinkedHashMap()
-            schema_properties.each { innerkey, value ->
-                ungrouped_params.put(innerkey, value)
-            }
-            paramsMap.put("Other parameters", ungrouped_params)
-        }
-
-        return paramsMap
-    }
-
-    //
-    // Get maximum number of characters across all parameter names
-    //
-    private static Integer paramsMaxChars( Map paramsMap) {
-        Integer maxChars = 0
-        for (group in paramsMap.keySet()) {
-            def Map group_params = (Map) paramsMap.get(group)  // This gets the parameters of that particular group
-            for (String param in group_params.keySet()) {
-                if (param.size() > maxChars) {
-                    maxChars = param.size()
-                }
-            }
-        }
-        return maxChars
     }
 }