+++ plugins/nf-schema/src/main/nextflow/validation/Utils.groovy
@@ -123,7 +123,7 @@ public class Utils {
     }
 
     // Resolve Schema path relative to main workflow directory
-    public static String getSchemaPath(String baseDir, String schemaFilename='nextflow_schema.json') {
+    public static String getSchemaPath(String baseDir, String schemaFilename) {
         if (Path.of(schemaFilename).exists()) {
             return schemaFilename
         } else {
@@ -268,5 +268,100 @@ public class Utils {
         colorcodes['biwhite']    = monochrome_logs ? '' : "\033[1;97m"
 
         return colorcodes
+    }
+
+    //
+    // This function tries to read a JSON params file
+    //
+    public static Map paramsLoad(Path json_schema) {
+        def paramsMap = [:]
+        try {
+            paramsMap = paramsRead(json_schema)
+        } catch (Exception e) {
+            println "Could not read parameters settings from JSON. $e"
+        }
+        return paramsMap
+    }
+
+    //
+    // Method to actually read in JSON file using Groovy.
+    // Group (as Key), values are all parameters
+    //    - Parameter1 as Key, Description as Value
+    //    - Parameter2 as Key, Description as Value
+    //    ....
+    // Group
+    //    -
+    private static Map paramsRead(Path json_schema) throws Exception {
+        def slurper = new JsonSlurper()
+        def Map schema = (Map) slurper.parse( json_schema )
+        // $defs is the adviced keyword for definitions. Keeping defs in for backwards compatibility
+        def Map schema_defs = (Map) (schema.get('$defs') ?: schema.get("defs"))
+        def Map schema_properties = (Map) schema.get('properties')
+        /* Tree looks like this in nf-core schema
+        * $defs <- this is what the first get('$defs') gets us
+                group 1
+                    title
+                    description
+                        properties
+                        parameter 1
+                            type
+                            description
+                        parameter 2
+                            type
+                            description
+                group 2
+                    title
+                    description
+                        properties
+                        parameter 1
+                            type
+                            description
+        * properties <- parameters can also be ungrouped, outside of $defs
+                parameter 1
+                    type
+                    description
+        */
+
+        def paramsMap = [:]
+        // Grouped params
+        if (schema_defs) {
+            for (group in schema_defs) {
+                def Map group_property = (Map) group.value['properties'] // Gets the property object of the group
+                def String title = (String) group.value['title']
+                def sub_params = [:]
+                group_property.each { innerkey, value ->
+                    sub_params.put(innerkey, value)
+                }
+                paramsMap.put(title, sub_params)
+            }
+        }
+
+        // Ungrouped params
+        if (schema_properties) {
+            def ungrouped_params = [:]
+            schema_properties.each { innerkey, value ->
+                ungrouped_params.put(innerkey, value)
+            }
+            paramsMap.put("Other parameters", ungrouped_params)
+        }
+
+        return paramsMap
+    }
+
+    //
+    // Get maximum number of characters across all parameter names
+    //
+    public static Integer paramsMaxChars( Map paramsMap ) {
+        return Collections.max(paramsMap.collect { _, val -> 
+            def Map groupParams = val as Map
+            longestStringLength(groupParams.keySet() as List<String> )
+        })
+    }
+
+    //
+    // Get the size of the longest string value in a list of strings
+    //
+    public static Integer longestStringLength( List<String> strings ) {
+        return strings ? Collections.max(strings.collect { it.size() }) : 0
     }
 }